# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Participación Ciudadana Documentación API",
        "description": "Endpoints de la app Participación Ciudadana",
        "contact": {
            "email": "ciudadana@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/v1",
            "description": "Endpoints de la app Participación Ciudadana"
        }
    ],
    "paths": {
...
        "/areas/{id}": {
            "put": {
                "tags": [
                    "Areas"
                ],
                "summary": "Update a area",
                "description": "Returns updated area",
                "operationId": "App\\Http\\Controllers\\AreaController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Area id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Areas"
                ],
                "summary": "Delete a area",
                "description": "Returns Json response",
                "operationId": "App\\Http\\Controllers\\AreaController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Area id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/breakdown/{id}": {
            "get": {
                "tags": [
                    "Breakdown"
                ],
                "summary": "Get a breakdown by id",
                "description": "Returns the breakdown",
                "operationId": "App\\Http\\Controllers\\BreakdownController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Breakdown id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "The breakdown not be found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Breakdown"
                ],
                "summary": "Update a breakdown",
                "description": "Returns updated breakdown",
                "operationId": "App\\Http\\Controllers\\BreakdownController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Breakdown id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Breakdown"
                ],
                "summary": "Delete a breakdown",
                "description": "Returns Json response",
                "operationId": "App\\Http\\Controllers\\BreakdownController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Breakdown id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/category/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a category by id",
                "description": "Returns the category",
                "operationId": "App\\Http\\Controllers\\CategoryController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "The category not be found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "description": "Returns updated category",
                "operationId": "App\\Http\\Controllers\\CategoryController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "description": "Returns Json response",
                "operationId": "App\\Http\\Controllers\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/incidence/{id}": {
            "get": {
                "tags": [
                    "Incidences"
                ],
                "summary": "Get a incidence by id",
                "description": "Returns the incedence",
                "operationId": "App\\Http\\Controllers\\IncidenceController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incidence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "The incidence not be found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Incidences"
                ],
                "summary": "Update a incidence",
                "description": "Returns updated incidence",
                "operationId": "App\\Http\\Controllers\\IncidenceController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "incidence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Incidences"
                ],
                "summary": "Delete a incidence",
                "description": "Returns Json response",
                "operationId": "App\\Http\\Controllers\\IncidenceController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incidence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/enrollment/{id}": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get a enrollment by id",
                "description": "Returns the enrollment",
                "operationId": "App\\Http\\Controllers\\MatriculaController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enrollment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "The enrollment not be found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Update a enrollment",
                "description": "Returns updated enrollment",
                "operationId": "App\\Http\\Controllers\\MatriculaController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enrollment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Delete a enrollment",
                "description": "Returns Json response",
                "operationId": "App\\Http\\Controllers\\MatriculaController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enrollment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/notice": {
            "get": {
                "tags": [
                    "Notice"
                ],
                "summary": "Get list of notices",
                "description": "Returns list of notices",
                "operationId": "App\\Http\\Controllers\\NoticeController::index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notices"
                ],
                "summary": "Create a new notice",
                "description": "Returns created notice",
                "operationId": "App\\Http\\Controllers\\NoticeController::store",
                "parameters": [
                    {
                        "name": "request",
                        "in": "path",
                        "description": "request all data",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/notice/{id}": {
            "get": {
                "tags": [
                    "Notices"
                ],
                "summary": "Get a notice by id",
                "description": "Returns the notice",
                "operationId": "App\\Http\\Controllers\\NoticeController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notice id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "The notice not be found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Notices"
                ],
                "summary": "Update a notice",
                "description": "Returns updated notice",
                "operationId": "App\\Http\\Controllers\\NoticeController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "notice id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notices"
                ],
                "summary": "Delete a notice",
                "description": "Returns Json response",
                "operationId": "App\\Http\\Controllers\\NoticeController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notice id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/publiccenter/{id}": {
            "get": {
                "tags": [
                    "PublicCenter"
                ],
                "summary": "Get a Public Center by id",
                "description": "Returns the public center",
                "operationId": "App\\Http\\Controllers\\PublicCenterController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PublicCenter id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "The public center not be found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "PublicCenter"
                ],
                "summary": "Update a public center",
                "description": "Returns updated public center",
                "operationId": "App\\Http\\Controllers\\PublicCenterController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PublicCenter id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PublicCenter"
                ],
                "summary": "Delete a Public Center",
                "description": "Returns Json response",
                "operationId": "App\\Http\\Controllers\\PublicCenterController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PublicCenter id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
...
}